modelasaservice/fastapi/requirements.txt
matplotlib==3.5.2
numpy>=1.17.0
pandas==1.2.5
Pillow==9.3.0
pydantic==1.9.1
scikit_learn==1.1.3
scipy==1.6.2
fastapi>=0.68.2
xarray==0.20.1
python-multipart==0.0.5
